Feature: added some interaction with bot:
* 2 main commands: /images_to_pdf and /convert_document
* /convert_document:
	1) getting document
	2) offers 4 extensions to convert
	3) (fb2 doesn't work at all, others work at some measure)
	4) sometimes user have to send command twice for bot to
	react
* /images_to_pdf:
	1) needs fixing

+++ mysite/tgbot/commands_service/tgbot_service.py
@@ -1,10 +1,9 @@
-import codecs
 import os
-
-import textract
+from pprint import pprint
 
 from . import tgbot_answers, extract_data, tgbot_file_service
 import enum
+import json
 
 
 Answers = tgbot_answers.Answers_class
@@ -20,6 +19,13 @@ class Message_Type(enum.Enum):
     garbage = 5
 
 
+class Extensions(enum.Enum):
+    pdf = 1
+    doc = 2
+    txt = 3
+    fb2 = 4
+
+
 def detect_message_type(request_body):
     if 'text' in request_body['message']:
         return Message_Type.text
@@ -33,13 +39,8 @@ def detect_message_type(request_body):
         return Message_Type.garbage
 
 
-def process_document(request_body):
-    chat_id = DataExtractor.get_chat_id(request_body)
-
-    Answers.send_message(chat_id, "You've sent a document")
-    file_name = DataExtractor.get_file_name(request_body)
-    file_id = DataExtractor.get_file_id(request_body)
-    new_extension = '.pdf'
+def process_document(file_name, file_id, chat_id, new_extension: Extensions):
+    new_extension = f'.{new_extension.name}'
 
     try:
         file_path = FileService.download_document(file_id, file_name)
@@ -50,29 +51,95 @@ def process_document(request_body):
         new_file_name = os.path.splitext(file_name)[0] + new_extension
         Answers.send_document(chat_id, new_file_name, zipObj)
 
+        print('file_path = ' + file_path)
+        print('zip_path = ' + zip_path)
+        print('new_file_path = ' + new_file_path)
+
         os.remove(file_path)
         zipObj.close()
         os.remove(zip_path)
-        os.remove(new_file_path)
+
+        if file_path != new_file_path:
+            os.remove(new_file_path)
 
     except Exception as e:
-        Answers.send_message(chat_id, f'{str(e)}')
+        print(f'{str(e)}')
+        Answers.send_message(chat_id, 'Sorry, some error occured :(')
+
+
+def save_to_json(request_body):
+    with open('mysite\\tgbot\\commands_service\\downloads\\data.json', 'r') as rd:
+        data_list = json.load(rd)
+
+    data_list.append(request_body)
+
+    with open('mysite\\tgbot\\commands_service\\downloads\\data.json', 'w') as fp:
+        json.dump(data_list, fp, indent=4)
+
+
+def find_last_command():
+    with open('mysite\\tgbot\\commands_service\\downloads\\data.json', 'r') as rd:
+        data_list: list = json.load(rd)
+        data_list.reverse()
+        for i in range(1, len(data_list)):
+            if detect_message_type(data_list[i]) is Message_Type.text:
+                return data_list[i]['message']['text']
+    return 'ERROR'
+
+
+
+def find_last_document():
+    with open('mysite\\tgbot\\commands_service\\downloads\\data.json', 'r') as rd:
+        data_list: list = json.load(rd)
+        data_list.reverse()
+
+        for i in range(0, len(data_list)):
+            if detect_message_type(data_list[i]) == Message_Type.document:
+                return data_list[i]['message']['document']['file_name'], data_list[i]['message']['document']['file_id']
+
+    return -1
 
 
 def execute_command(request_body):
     chat_id = DataExtractor.get_chat_id(request_body)
+    save_to_json(request_body)
 
     if detect_message_type(request_body) is Message_Type.text:
-        Answers.send_message(chat_id, "You've sent a text:\n" + f'{DataExtractor.get_message_text(request_body)}')
+
+        file_name, file_id = find_last_document()
+
+        if find_last_command() == f'/{Extensions.pdf.name}':
+            Answers.send_message(chat_id, 'Processing document. New extension - .pdf')
+            process_document(file_name, file_id, chat_id, Extensions.pdf)
+
+        elif find_last_command() == f'/{Extensions.doc.name}':
+            Answers.send_message(chat_id, 'Processing document. New extension - .doc')
+            process_document(file_name, file_id, chat_id, Extensions.doc)
+
+        elif find_last_command() == f'/{Extensions.txt.name}':
+            Answers.send_message(chat_id, 'Processing document. New extension - .txt')
+            process_document(file_name, file_id, chat_id, Extensions.txt)
+
+        elif find_last_command() == f'/{Extensions.fb2.name}':
+            Answers.send_message(chat_id, 'Processing document. New extension - .fb2')
+            process_document(file_name, file_id, chat_id, Extensions.fb2)
+
+
+        elif DataExtractor.get_message_text(request_body) == '/images_to_pdf':
+            Answers.send_message(chat_id, 'Waiting for images...')
+
+        elif DataExtractor.get_message_text(request_body) == '/convert_document':
+            Answers.send_message(chat_id, 'Waiting for document...')
 
     elif detect_message_type(request_body) is Message_Type.document:
-        process_document(request_body)
+        Answers.send_message(chat_id, "Choose 1 of the following extensions:")
+        Answers.send_message(chat_id, "/pdf\n/doc\n/txt\n/fb2 (this one I need to fix)")
 
     elif detect_message_type(request_body) is Message_Type.image:
-        Answers.send_message(chat_id, "You've sent an image")
+        if find_last_command() == '/images_to_pdf':
+            # Треба завантажити всі фото починаючи з останньої команди /images_to_pdf
+            Answers.send_message(chat_id, 'Creating pdf. Wait a second 🙃')
 
-    elif detect_message_type(request_body) is Message_Type.sticker:
-        Answers.send_message(chat_id, "You've sent a sticker")
 
-    elif detect_message_type(request_body) is Message_Type.garbage:
-        Answers.send_message(chat_id, "Cannot detect the type of file :(")
+    else:
+        Answers.send_message(chat_id, "Oops(")
